// src/types/IStanding.ts
export interface IStanding {
  awayDraws: number;
  awayLosses: number;
  awayWins: number;
  championship: string;
  cupName: string;
  division: string;
  draws: number;
  efficiency: number;
  gamesPlayed: number;
  goalsAgainst: number;
  goalsDifference: number;
  goalsFor: number;
  groupname: string;
  homeDraws: number;
  homeLosses: number;
  homeWins: number;
  id: number;
  idChampionship: number;
  idTeam: number;
  losses: number;
  maxPointsPossible: number;
  phase: string;
  points: number;
  position: number;
  redCards: number;
  round: number;
  team: string;
  wins: number;
  yellowCards: number;
}